fix(lsp): sync tsx vue dependency shims
